refactor: Create index.js for grouping and exporting all group-related components
