+++ frontend/src/App.js
@@ -3,9 +3,9 @@ import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
 import { ToastContainer } from 'react-toastify';
 import 'react-toastify/dist/ReactToastify.css';
 import { API_ENDPOINTS } from './config/api';
-import Headers from './components/Headers';
-import Footer from './components/Footer';
-import Modal from './components/Modal';
+import Headers from './components/common/Headers';
+import Footer from './components/common/Footer';
+import Modal from './components/common/Modal';
 import HomePage from './pages/HomePage';
 import Inicio from './pages/Inicio';
 import ComoFunciona from './pages/ComoFunciona';
