+++ frontend/src/components/common/Footer.css
