+++ frontend/src/components/common/Hero.css
