+++ frontend/src/components/common/index.js
@@ -1,0 +1,6 @@
+// Archivo barrel para exportar todos los componentes comunes
+export { default as Headers } from './Headers';
+export { default as Footer } from './Footer';
+export { default as Hero } from './Hero';
+export { default as Features } from './Features';
+export { default as Modal } from './Modal';
