+++ frontend/src/components/grupos/DetalleGrupo/DetalleGrupo.css
