+++ frontend/src/components/grupos/EliminarGrupoModal/EliminarGrupoModal.js
@@ -2,7 +2,7 @@ import React, { useState } from 'react';
 import { toast } from 'react-toastify';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faExclamationTriangle } from '@fortawesome/free-solid-svg-icons';
-import { API_ENDPOINTS } from '../config/api';
+import { API_ENDPOINTS } from '../../../config/api';
 import './EliminarGrupoModal.css';
 
 const EliminarGrupoModal = ({ isOpen, onClose, grupo, onGrupoEliminado }) => {
