+++ frontend/src/components/grupos/ParticipantCard/ParticipantCard.css
