+++ frontend/src/components/grupos/index.js
@@ -1,0 +1,7 @@
+// Archivo barrel para exportar todos los componentes de grupos
+export { default as DetalleGrupo } from './DetalleGrupo/DetalleGrupo';
+export { default as CrearGrupoForm } from './CrearGrupoForm/CrearGrupoForm';
+export { default as EditarGrupoForm } from './EditarGrupoForm/EditarGrupoForm';
+export { default as EliminarGrupoModal } from './EliminarGrupoModal/EliminarGrupoModal';
+export { default as UnirseGrupoForm } from './UnirseGrupoForm/UnirseGrupoForm';
+export { default as ParticipantCard } from './ParticipantCard/ParticipantCard';
